requirements.txt
boto>=2.36.0
pytz==2014.10
python-dateutil>=2.1,<3.0.0
XlsxWriter==0.6.7
PyYAML>=3.11
nose==1.3.4

